可以正常使用spark－submit工具进行提交，编写脚本nohup &可以后台运行。大多数的功能已经实现。经过简单测试，存在以下问题： 一、Filter 1. DTime不起作用 2. SampleInterval不起作用 3. GPSRange filter没有实现 4. PassbyArea没有出现预期效果 二、Feature 1. SamplePointSpeed没有实现 2. Duration的显示，横坐标需要降低精度。

+++ src/main/scala/cn/edu/suda/ada/spatialspark/features/NumbericalClassifier.scala
@@ -12,8 +12,8 @@ import cn.edu.suda.ada.spatialspark.core.{GPSPoint, Trajectory}
  */
 
 trait TrajectoryNumericalClassifier{
-  var levelStep: Long = 1
-  def setLevelStep(levelStep : Long): Unit ={
+  var levelStep: Int = 1
+  def setLevelStep(levelStep : Int): Unit ={
     this.levelStep = levelStep
   }
   /**
